Move the mock api server to a separate module
Tests directories (where the mock server was until now) are in .dockerignore,
so it would be imposible to use the mock server in a docker compose setup.
